refactor: useEffect chart 바뀐 값 있을때만 setState 호출
